Heavy refactor. Clarify file and code location
