--- access/magic-retrieve-link.php
@@ -1,197 +1,0 @@
-<?php
-if ( !defined( 'ABSPATH' ) ) { exit; } // Exit if accessed directly.
-
-class DT_Streams_Retrieve_Link extends DT_Magic_Url_Base
-{
-    public $magic = false;
-    public $parts = false;
-    public $page_title = 'Retrieve Link';
-    public $root = 'streams_app';
-    public $type = 'retrieve_link';
-    public $type_name = 'Retrieve Link';
-    public static $token = 'streams_app_retrieve_link';
-    public $post_type = 'streams'; // This can be supplied or not supplied. It does not influence the url verification.
-
-    private static $_instance = null;
-    public static function instance() {
-        if ( is_null( self::$_instance ) ) {
-            self::$_instance = new self();
-        }
-        return self::$_instance;
-    } // End instance()
-
-    public function __construct() {
-        parent::__construct();
-
-        $url = dt_get_url_path();
-        if ( ( $this->root . '/' . $this->type ) === $url ) {
-
-            $this->magic = new DT_Magic_URL( $this->root );
-            $this->parts = $this->magic->parse_url_parts();
-
-
-            // register url and access
-            add_action( 'template_redirect', [ $this, 'theme_redirect' ] );
-            add_filter( 'dt_blank_access', function (){ return true;
-            }, 100, 1 );
-            add_filter( 'dt_allow_non_login_access', function (){ return true;
-            }, 100, 1 );
-            add_filter( 'dt_override_header_meta', function (){ return true;
-            }, 100, 1 );
-
-            // header content
-            add_filter( 'dt_blank_title', [ $this, 'page_tab_title' ] ); // adds basic title to browser tab
-            add_action( 'wp_print_scripts', [ $this, 'print_scripts' ], 1500 ); // authorizes scripts
-            add_action( 'wp_print_styles', [ $this, 'print_styles' ], 1500 ); // authorizes styles
-
-
-            // page content
-            add_action( 'dt_blank_head', [ $this, '_header' ] );
-            add_action( 'dt_blank_footer', [ $this, '_footer' ] );
-            add_action( 'dt_blank_body', [ $this, 'body' ] ); // body for no post key
-
-            add_filter( 'dt_magic_url_base_allowed_css', [ $this, 'dt_magic_url_base_allowed_css' ], 10, 1 );
-            add_filter( 'dt_magic_url_base_allowed_js', [ $this, 'dt_magic_url_base_allowed_js' ], 10, 1 );
-            add_action( 'wp_enqueue_scripts', [ $this, '_wp_enqueue_scripts' ], 100 );
-        }
-
-        if ( dt_is_rest() ) {
-            add_action( 'rest_api_init', [ $this, 'add_endpoints' ] );
-            add_filter( 'dt_allow_rest_access', [ $this, 'authorize_url' ], 10, 1 );
-        }
-    }
-
-    public function dt_magic_url_base_allowed_js( $allowed_js ) {
-        return $allowed_js;
-    }
-
-    public function dt_magic_url_base_allowed_css( $allowed_css ) {
-        return $allowed_css;
-    }
-
-    public function header_style(){
-        ?>
-        <style>
-            body {
-                background-color:white;
-            }
-            #content {
-                margin-left: auto;
-                margin-right: auto;
-                max-width: 1440px;
-                font-size: 5rem;
-                margin-top: 30%;
-                text-align: center;
-            }
-        </style>
-        <?php
-    }
-
-    public function body(){
-        ?>
-        <div id="content"><span class="first"><i class="loading-spinner active"></i></span><span class="second"></span></div>
-        <?php
-    }
-
-    public function footer_javascript(){
-        ?>
-        <script>
-            console.log('insert footer_javascript')
-
-            let jsObject = [<?php echo json_encode([
-                'map_key' => DT_Mapbox_API::get_key(),
-                'root' => esc_url_raw( rest_url() ),
-                'nonce' => wp_create_nonce( 'wp_rest' ),
-                'parts' => $this->parts,
-                'translations' => [
-                    'add' => __( 'Add Magic', 'disciple-tools-plugin-starter-template' ),
-                ],
-            ]) ?>][0]
-
-            window.get_magic = () => {
-                jQuery.ajax({
-                    type: "POST",
-                    data: JSON.stringify( { action: 'get', parts: jsObject.parts } ),
-                    contentType: "application/json; charset=utf-8",
-                    dataType: "json",
-                    url: jsObject.root + jsObject.parts.root + '/v1/' + jsObject.parts.type,
-                    beforeSend: function (xhr) {
-                        xhr.setRequestHeader('X-WP-Nonce', jsObject.nonce )
-                    }
-                })
-                    .done(function(data){
-                        jQuery('#content .first').html('Insert Your Page')
-                    })
-                    .fail(function(e) {
-                        console.log(e)
-                        jQuery('#error').html(e)
-                    })
-
-                jQuery.ajax({
-                    type: "POST",
-                    data: JSON.stringify( { action: 'excited', parts: jsObject.parts } ),
-                    contentType: "application/json; charset=utf-8",
-                    dataType: "json",
-                    url: jsObject.root + jsObject.parts.root + '/v1/' + jsObject.parts.type,
-                    beforeSend: function (xhr) {
-                        xhr.setRequestHeader('X-WP-Nonce', jsObject.nonce )
-                    }
-                })
-                    .done(function(data){
-                        jQuery('#content .second').html('!!')
-                    })
-                    .fail(function(e) {
-                        console.log(e)
-                        jQuery('#error').html(e)
-                    })
-            }
-            window.get_magic()
-
-        </script>
-        <?php
-        return true;
-    }
-
-    public static function _wp_enqueue_scripts(){
-    }
-
-    /**
-     * Register REST Endpoints
-     * @link https://github.com/DiscipleTools/disciple-tools-theme/wiki/Site-to-Site-Link for outside of wordpress authentication
-     */
-    public function add_endpoints() {
-        $namespace = $this->root . '/v1';
-        register_rest_route(
-            $namespace,
-            '/'.$this->type,
-            [
-                [
-                    'methods'  => WP_REST_Server::CREATABLE,
-                    'callback' => [ $this, 'endpoint' ],
-                ],
-            ]
-        );
-    }
-
-    public function endpoint( WP_REST_Request $request ) {
-        $params = $request->get_params();
-
-        if ( ! isset( $params['parts'], $params['action'] ) ) {
-            return new WP_Error( __METHOD__, 'Missing parameters', [ 'status' => 400 ] );
-        }
-
-        $params = dt_recursive_sanitize_array( $params );
-
-        switch ( $params['action'] ) {
-            case 'get':
-                // do something
-                return true;
-            case 'excited':
-                // do something else
-            default:
-                return true;
-        }
-    }
-
-}
-DT_Streams_Retrieve_Link::instance();
