+++ magic/streams-self-register.php
@@ -9,6 +9,7 @@ class DT_Streams_Self_Register extends DT_Magic_Url_Self_Register {
     public $portal_url = 'streams_app/report/';
     public $type_name = 'Access';
     public $post_type = 'streams';
+    public $page_description = '';
 
     private static $_instance = null;
     public static function instance() {
@@ -22,8 +23,6 @@ class DT_Streams_Self_Register extends DT_Magic_Url_Self_Register {
         $this->meta_key = $this->root . '_' . $this->type . '_magic_key';
         parent::__construct();
 
-
-
         // fail if not valid url
         $url = dt_get_url_path();
         if ( strpos( $url, $this->root . '/' . $this->type ) === false ) {
@@ -35,18 +34,18 @@ class DT_Streams_Self_Register extends DT_Magic_Url_Self_Register {
         }
 
         add_action( 'dt_blank_body', [ $this, 'body' ] );
-        add_filter( 'dt_magic_url_base_allowed_css', [ $this, 'dt_magic_url_base_allowed_css' ], 10, 1 );
-        add_filter( 'dt_magic_url_base_allowed_js', [ $this, 'dt_magic_url_base_allowed_js' ], 10, 1 );
+//        add_filter( 'dt_magic_url_base_allowed_css', [ $this, 'dt_magic_url_base_allowed_css' ], 10, 1 );
+//        add_filter( 'dt_magic_url_base_allowed_js', [ $this, 'dt_magic_url_base_allowed_js' ], 10, 1 );
 
     }
 
-    public function dt_magic_url_base_allowed_js( $allowed_js ) {
-        return $allowed_js;
-    }
-
-    public function dt_magic_url_base_allowed_css( $allowed_css ) {
-        return $allowed_css;
-    }
+//    public function dt_magic_url_base_allowed_js( $allowed_js ) {
+//        return $allowed_js;
+//    }
+//
+//    public function dt_magic_url_base_allowed_css( $allowed_css ) {
+//        return $allowed_css;
+//    }
 
     public function header_javascript(){
         $this->register_form_header_javascript();
